uanfilms_backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
gunicorn==23.0.0
mysql-connector-python==9.4.0
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
django-storages[azure]==1.14.2
azure-identity==1.17.1
Pillow==10.1.0 
PyJWT==2.10.1
python-dotenv==1.2.1